requirements.txt
beautifulsoup4==4.14.2
fastapi==0.121.3
pillow==12.0.0
pydantic==2.12.4
python-multipart==0.0.20
pyyaml==6.0.3
uvicorn==0.38.0
websockets==15.0.1
